requirements.txt
faker
kafka-python
pyspark
pandas